hops

+++ main/views.py
@@ -1,45 +1,67 @@
 from django.shortcuts import render,redirect,HttpResponse
 from . import data as db
 from django.contrib import messages
+import os
 
+# from dotenv import load_dotenv
+# load_dotenv()
 
+def top():
+    data=db.get_doc()
+    d=[]
+    for i in range(3):
+        try:
+            d.append(data[i])
+        except:
+            break
+        
+    data={'data':d}
+    return data
 
+def notifination(request,key=""):
 
-
-def notifination(request):
-    if request.method=='POST':
-        id=request.POST.get('id')
-        db.delete_doc(id)
+    
+    if key==os.getenv('NOT_KEY'):
+        if request.method=='POST':
+            id=request.POST.get('id')
+            db.delete_doc(id)
+        
+        data=db.get_doc()
+        data={'data':data,"key":os.getenv('NOT_KEY')}
+        return render(request, 'admin.html',data)
     
     data=db.get_doc()
-    data={'data':data}
-    return render(request, 'admin.html',data)
+    data=top()
+    return render(request, 'home.html',data)
 
+def feedbacks(request):
+    data=db.get_doc()
+    data={'data':data}
+    return render(request, 'feedback.html',data)
 
 def home(request):
     if request.method == 'POST':
         name=request.POST.get('name')
-        phno=request.POST.get('phno')
-        date_in=request.POST.get('date_in')
-        date_out=request.POST.get('date_out')
-        people=request.POST.get('people')
+        # phno=request.POST.get('phno')
+        # date_in=request.POST.get('date_in')
+        # date_out=request.POST.get('date_out')
+        Ratings=request.POST.get('Ratings')
         msg=request.POST.get('msg')
-        place=request.POST.get('place')
+        # place=request.POST.get('place')
         
-        print(type(date_in))
-        if name==""or phno=="" or date_in=="" or date_out=="" or people=="" or place=="" or date_in>date_out:
+        # print(type(date_in))
+        if name=="":
             messages.error(request, 'Somthing Went Wrong')
             return redirect('home')  # Redirect to the same page
         
-        data = {"Name": name ,"Phone_No":phno,
-        "Check_In": date_in, "Check_out": date_out, "No-of-Members": people, "Custom_msg": msg,"Place":place}
+        data = {"Name": name , "Ratings": Ratings, "Custom_msg": msg}
         
        
         
         
         result=db.add_doc(data)
         if result:
-            messages.success(request, 'Thank You! Contact For Booking Conformation')
+            messages.success(request, 'Thank You For Responding!')
             return redirect('home')  # Redirect to the same page
         messages.success(request, 'Somthing Went Wrong')
         return redirect('home')  # Redirect to the same page
@@ -48,4 +70,6 @@ def home(request):
        
    
     
-    return render(request, 'home.html')
+        
+    data=top()
+    return render(request, 'home.html',data)
